enhance condense metric
This PR supports cases where there are dimensions which only act as filters, and don't have an "All" level. In such cases, condense_metric usually returns NAs as it does not find "All". The fix here checks if there are such dimensions, and if so, uses a group, nest, and mutate strategy to condense the metric.